COSI-72: add-mock-package-and-use-s3-common-mock
